fix: correctly read the ispmd of a Sprm
The correct way to get the ispmd is to AND the sprm with 0x1ff according
to section 2.6 of [MS-DOC].

The current buggy behavior caused some ismpd-s (e.g. 0x40) to be parsed
as 0, which was then treated as a sprmCFRMarkDel and caused text to be
deleted.

I also changed all right shifts (>>) into unsigned right shifts (>>>),
since all the shifted values are unsigned.